packages/launcher/pyproject.toml
[project]
name = "zagent-launcher"
version = "0.1.0"
description = "Host-side CLI and Python API for launching ZAgent runtime containers."
readme = "README.md"
authors = [
    { name = "vfaddey", email = "vfaddey@yandex.ru" },
]
requires-python = ">=3.12,<3.14"
dependencies = [
    "dishka>=1.6,<2",
    "docker>=7.1,<8",
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "rich>=13.9",
    "typer>=0.15",
]

[project.scripts]
zagent = "zagent_launcher.presentation.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zagent_launcher"]
